requirements.txt
requests
faiss-cpu
langchain
pandas
streamlit
openai